create common package for shared structs
