+++ client/mcp_tool.go
@@ -4,33 +4,13 @@ import (
 	"bytes"
 	"encoding/json"
 	"fmt"
+	"github.com/mcpjungle/mcpjungle/pkg/types"
 	"io"
 	"net/http"
 )
 
-// ToolInputSchema defines the schema for the input parameters of a tool
-type ToolInputSchema struct {
-	Type       string         `json:"type"`
-	Properties map[string]any `json:"properties,omitempty"`
-	Required   []string       `json:"required,omitempty"`
-}
-
-// Tool represents a tool provided by an MCP Server registered in the registry.
-type Tool struct {
-	Name        string          `json:"name"`
-	Enabled     bool            `json:"enabled"`
-	Description string          `json:"description"`
-	InputSchema ToolInputSchema `json:"input_schema"`
-}
-
-type ToolInvokeResult struct {
-	Meta    map[string]any   `json:"_meta,omitempty"`
-	IsError bool             `json:"isError,omitempty"`
-	Content []map[string]any `json:"content"`
-}
-
 // ListTools fetches the list of tools, optionally filtered by server name.
-func (c *Client) ListTools(server string) ([]*Tool, error) {
+func (c *Client) ListTools(server string) ([]*types.Tool, error) {
 	u, _ := c.constructAPIEndpoint("/tools")
 	req, _ := c.newRequest(http.MethodGet, u, nil)
 	if server != "" {
@@ -49,7 +29,7 @@ func (c *Client) ListTools(server string) ([]*Tool, error) {
 		return nil, fmt.Errorf("request failed with status: %d, message: %s", resp.StatusCode, body)
 	}
 
-	var tools []*Tool
+	var tools []*types.Tool
 	if err := json.NewDecoder(resp.Body).Decode(&tools); err != nil {
 		return nil, fmt.Errorf("failed to decode response: %w", err)
 	}
@@ -117,7 +97,7 @@ func (c *Client) DisableTools(name string) ([]string, error) {
 }
 
 // GetTool fetches a specific tool by its name.
-func (c *Client) GetTool(name string) (*Tool, error) {
+func (c *Client) GetTool(name string) (*types.Tool, error) {
 	u, _ := c.constructAPIEndpoint("/tool")
 	req, _ := c.newRequest(http.MethodGet, u, nil)
 	q := req.URL.Query()
@@ -135,7 +115,7 @@ func (c *Client) GetTool(name string) (*Tool, error) {
 		return nil, fmt.Errorf("request failed with status: %d, message: %s", resp.StatusCode, body)
 	}
 
-	var tool Tool
+	var tool types.Tool
 	if err := json.NewDecoder(resp.Body).Decode(&tool); err != nil {
 		return nil, fmt.Errorf("failed to decode response: %w", err)
 	}
@@ -144,7 +124,7 @@ func (c *Client) GetTool(name string) (*Tool, error) {
 
 // InvokeTool sends a JSON payload to invoke a tool.
 // For now, this function only supports invoking tools that return a string response.
-func (c *Client) InvokeTool(name string, input map[string]any) (*ToolInvokeResult, error) {
+func (c *Client) InvokeTool(name string, input map[string]any) (*types.ToolInvokeResult, error) {
 	// We need to insert the tool name into the POST payload
 	// In order not to mutate the user-supplied input, create a shallow copy of the input
 	// and add the name field to it.
@@ -174,7 +154,7 @@ func (c *Client) InvokeTool(name string, input map[string]any) (*ToolInvokeResul
 		return nil, fmt.Errorf("request failed with status: %d, message: %s", resp.StatusCode, string(respBody))
 	}
 
-	var result *ToolInvokeResult
+	var result *types.ToolInvokeResult
 	if err := json.Unmarshal(respBody, &result); err != nil {
 		return nil, fmt.Errorf("failed to decode response: %w", err)
 	}
