+++ pkg/types/doc.go
@@ -1,0 +1,7 @@
+// Package types provides shared data structures used by both the mcpjungle server
+// and client components. These structures define the contract between the
+// server and client and should be used for data exchange.
+package types
+
+// TODO: refactor the http server handlers to use these types rather than the
+//  gorm models when interacting with clients
